requirements.txt
python-telegram-bot==21.6
langchain==0.3.4
langchain-community>=0.3.3
langchain-openai==0.2.1
langgraph==0.2.34
python-dotenv==1.0.1
openai==1.51.0
bs4==0.0.2
tiktoken==0.7.0
chromadb==0.5.11
langchain-chroma==0.1.4
langchain-ollama==0.2.0
ollama==0.3.3
llm-adaptive-router==0.1.13
langchain-anthropic==0.2.3
python-multipart
markdown
colorama
beautifulsoup4
colorama
md2pdf
arxiv
PyMuPDF
requests
jinja2
aiofiles
mistune
python-docx
htmldocx
lxml_html_clean
websockets
unstructured
json_repair
json5
loguru
deep-translator==1.11.4